Removed the .gitattribute file
Although I'm not sure why it must be removed, it certainly doesn't hurt anything. See https://git-scm.com/docs/gitattributes The "text" attribute enables and controls end-of-line normalization. When text is set to "auto", the path is marked for automatic end-of-line conversion. If Git decides that the content is text, its line endings are converted to LF on checkin. When the file has been committed with CRLF, no conversion is done.